fix typo in public method names, provide aliases for backward compatibility
